# [#5987] Unable to initialize PlatformDependent when sun.misc.Unsafe i…