@Scottmitch 2016-11-08
This change was introduced in 
https://github.com/netty/netty/commit/c2de195f872506dacd5df5cd5507b49e1669790d for performance reasons. Can we please re-run the same benchmarks and report results (before/after)?

/cc @buchgr