@johnou 2016-11-08
Before

Benchmark                                    (size)  Mode  Cnt      Score     Error  Units
PlatformDependentBenchmark.arraysBytesEqual      10  avgt   20      7.841 ±   0.192  ns/op
PlatformDependentBenchmark.arraysBytesEqual      50  avgt   20     20.704 ±   0.165  ns/op
PlatformDependentBenchmark.arraysBytesEqual     100  avgt   20     36.425 ±   0.726  ns/op
PlatformDependentBenchmark.arraysBytesEqual    1000  avgt   20    319.758 ±   4.208  ns/op
PlatformDependentBenchmark.arraysBytesEqual   10000  avgt   20   3122.020 ±  64.212  ns/op
PlatformDependentBenchmark.arraysBytesEqual  100000  avgt   20  29962.546 ± 373.961  ns/op
PlatformDependentBenchmark.unsafeBytesEqual      10  avgt   20      4.724 ±   0.052  ns/op
PlatformDependentBenchmark.unsafeBytesEqual      50  avgt   20      8.826 ±   0.116  ns/op
PlatformDependentBenchmark.unsafeBytesEqual     100  avgt   20     12.603 ±   0.142  ns/op
PlatformDependentBenchmark.unsafeBytesEqual    1000  avgt   20     78.768 ±   1.020  ns/op
PlatformDependentBenchmark.unsafeBytesEqual   10000  avgt   20    766.275 ±  70.988  ns/op
PlatformDependentBenchmark.unsafeBytesEqual  100000  avgt   20   7378.607 ±  82.921  ns/op

After

Benchmark                                    (size)  Mode  Cnt      Score     Error  Units
PlatformDependentBenchmark.arraysBytesEqual      10  avgt   20      7.977 ±   0.074  ns/op
PlatformDependentBenchmark.arraysBytesEqual      50  avgt   20     21.471 ±   0.299  ns/op
PlatformDependentBenchmark.arraysBytesEqual     100  avgt   20     37.157 ±   0.521  ns/op
PlatformDependentBenchmark.arraysBytesEqual    1000  avgt   20    325.312 ±   7.417  ns/op
PlatformDependentBenchmark.arraysBytesEqual   10000  avgt   20   3162.007 ±  67.813  ns/op
PlatformDependentBenchmark.arraysBytesEqual  100000  avgt   20  30257.358 ± 403.784  ns/op
PlatformDependentBenchmark.unsafeBytesEqual      10  avgt   20      4.858 ±   0.072  ns/op
PlatformDependentBenchmark.unsafeBytesEqual      50  avgt   20      9.336 ±   0.607  ns/op
PlatformDependentBenchmark.unsafeBytesEqual     100  avgt   20     12.679 ±   0.123  ns/op
PlatformDependentBenchmark.unsafeBytesEqual    1000  avgt   20     79.441 ±   0.748  ns/op
PlatformDependentBenchmark.unsafeBytesEqual   10000  avgt   20    716.150 ±   7.341  ns/op
PlatformDependentBenchmark.unsafeBytesEqual  100000  avgt   20   7315.621 ±  74.567  ns/op