@johnou 2016-11-08
@Scottmitch I am a little confused at how this change could affect performance, it's a constant after all, right? Is there some inline magic going on? What about using a static block instead..?

```java
    private static final long BYTE_ARRAY_BASE_OFFSET;
    
    static {
        BYTE_ARRAY_BASE_OFFSET = !hasUnsafe() ? -1 : PlatformDependent0.byteArrayBaseOffset();
    }
```