@johnou 2016-11-08
After

ObserverBenchmark                                    (size)  Mode  Cnt      Score     Error  Units
PlatformDependentBenchmark.arraysBytesEqual      10  avgt   20      7.977 ±   0.074  ns/op
PlatformDependentBenchmark.arraysBytesEqual      50  avgt   20     21.471 ±   0.299  ns/op
PlatformDependentBenchmark.arraysBytesEqual     100  avgt   20     37.157 ±   0.521  ns/op
PlatformDependentBenchmark.arraysBytesEqual    1000  avgt   20    325.312 ±   7.417  ns/op
PlatformDependentBenchmark.arraysBytesEqual   10000  avgt   20   3162.007 ±  67.813  ns/op
PlatformDependentBenchmark.arraysBytesEqual  100000  avgt   20  30257.358 ± 403.784  ns/op
PlatformDependentBenchmark.unsafeBytesEqual      10  avgt   20      4.858 ±   0.072  ns/op
PlatformDependentBenchmark.unsafeBytesEqual      50  avgt   20      9.336 ±   0.607  ns/op
PlatformDependentBenchmark.unsafeBytesEqual     100  avgt   20     12.679 ±   0.123  ns/op
PlatformDependentBenchmark.unsafeBytesEqual    1000  avgt   20     79.441 ±   0.748  ns/op
PlatformDependentBenchmark.unsafeBytesEqual   10000  avgt   20    716.150 ±   7.341  ns/op
PlatformDependentBenchmark.unsafeBytesEqual  100000  avgt   20   7315.621 ±  74.567  ns/op

Static block

Benchmark                                    (size)  Mode  Cnt      Score     Error  Units
PlatformDependentBenchmark.arraysBytesEqual      10  avgt   20      8.367 ±   0.092  ns/op
PlatformDependentBenchmark.arraysBytesEqual      50  avgt   20     22.276 ±   0.255  ns/op
PlatformDependentBenchmark.arraysBytesEqual     100  avgt   20     38.225 ±   0.696  ns/op
PlatformDependentBenchmark.arraysBytesEqual    1000  avgt   20    334.379 ±   2.878  ns/op
PlatformDependentBenchmark.arraysBytesEqual   10000  avgt   20   3256.167 ±  38.907  ns/op
PlatformDependentBenchmark.arraysBytesEqual  100000  avgt   20  31253.648 ± 246.035  ns/op
PlatformDependentBenchmark.unsafeBytesEqual      10  avgt   20      5.030 ±   0.073  ns/op
PlatformDependentBenchmark.unsafeBytesEqual      50  avgt   20      9.245 ±   0.081  ns/op
PlatformDependentBenchmark.unsafeBytesEqual     100  avgt   20     13.314 ±   0.268  ns/op
PlatformDependentBenchmark.unsafeBytesEqual    1000  avgt   20     83.415 ±   1.757  ns/op
PlatformDependentBenchmark.unsafeBytesEqual   10000  avgt   20    750.704 ±   8.623  ns/op
PlatformDependentBenchmark.unsafeBytesEqual  100000  avgt   20   7681.236 ±  82.916  ns/op