@normanmaurer 2016-11-08
@Scottmitch nevermind... Let me just merge this. This should make no difference as it is just called on "init time" anyway.