@normanmaurer 2016-11-08
Cherry-picked into 4.1 (895a92cb2208906e0cabba21063e604c10c4d5ae) . 4.0 is not affected. 

@johnou thanks!